*: move "PushMetric" to pkg/metricutil
